--- src/data_utils/create_dataset/dbconnection.json
@@ -1,8 +1,0 @@
-{
-    "dbname": "cgm",
-    "user": "azure_ml@ml-postgresql-sql-prod",
-    "host": "ml-postgresql-sql-prod.postgres.database.azure.com",
-    "password": "armumfang",
-    "port": 5432,
-    "sslmode": "require"
-}
