--- src/data_utils/create_dataset/parameters.yml
@@ -1,11 +1,0 @@
-database:
-    db_connection_file: 'dbconnection.json'
-scans:
-    scan_amount : 5000
-calibration:
-    calibration_file: '../../common/depthmap_toolkit/camera_calibration.txt'
-data:
-    scangroup: 'test' 
-paths:
-    source_path: '/mnt/bmzdata/cgminbmzprod_v5.0/'
-    target_path: 'evaluation/'
